refactor: improve tests
Take the same approach to the tests as the heater-shaker, condensing
host comms message checking.
