+++ stm32-modules/thermocycler-gen2/tests/CMakeLists.txt
@@ -22,6 +22,7 @@ add_executable(${TARGET_MODULE_NAME}
     test_colors.cpp
     test_motor_task.cpp
     test_motor_utils.cpp
+    test_motor_state_machine.cpp
     test_eeprom.cpp
     test_errors.cpp
     # GCode parse tests
