+++ stm32-modules/thermocycler-gen2/tests/test_lid_heater_task.cpp
@@ -31,17 +31,8 @@ SCENARIO("lid heater task message passing") {
             THEN("the task should get the message") {
                 REQUIRE(tasks->get_lid_heater_queue().backing_deque.empty());
                 AND_THEN("the task should respond to the messsage") {
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(std::holds_alternative<
-                            messages::GetLidTemperatureDebugResponse>(
-                        response));
-                    auto gettemp =
-                        std::get<messages::GetLidTemperatureDebugResponse>(
-                            response);
+                    auto gettemp = tasks->get_latest_host_comms_message<
+                        messages::GetLidTemperatureDebugResponse>();
                     REQUIRE(gettemp.responding_to_id == message.id);
                     REQUIRE_THAT(gettemp.lid_temp,
                                  Catch::Matchers::WithinAbs(_valid_temp, 0.1));
@@ -57,16 +48,8 @@ SCENARIO("lid heater task message passing") {
             THEN("the task should get the message") {
                 REQUIRE(tasks->get_lid_heater_queue().backing_deque.empty());
                 AND_THEN("the task should respond to the messsage") {
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(
-                        std::holds_alternative<messages::GetLidTempResponse>(
-                            response));
-                    auto gettemp =
-                        std::get<messages::GetLidTempResponse>(response);
+                    auto gettemp = tasks->get_latest_host_comms_message<
+                        messages::GetLidTempResponse>();
                     REQUIRE(gettemp.responding_to_id == message.id);
                     REQUIRE_THAT(gettemp.current_temp,
                                  Catch::Matchers::WithinAbs(_valid_temp, 0.1));
@@ -86,20 +69,7 @@ SCENARIO("lid heater task message passing") {
                 AND_THEN("the task should act on the message") {
                     REQUIRE(
                         tasks->get_thermal_plate_queue().backing_deque.empty());
-
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(
-                        std::holds_alternative<messages::AcknowledgePrevious>(
-                            response));
-                    auto response_msg =
-                        std::get<messages::AcknowledgePrevious>(response);
-                    REQUIRE(response_msg.responding_to_id == 123);
-                    REQUIRE(response_msg.with_error ==
-                            errors::ErrorCode::NO_ERROR);
+                    tasks->require_has_ack_for(message);
                     REQUIRE(tasks->get_lid_heater_policy().get_heater_power() ==
                             0.65);
                 }
@@ -120,20 +90,7 @@ SCENARIO("lid heater task message passing") {
                 AND_THEN("the task should act on the message") {
                     REQUIRE(
                         tasks->get_thermal_plate_queue().backing_deque.empty());
-
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(
-                        std::holds_alternative<messages::AcknowledgePrevious>(
-                            response));
-                    auto response_msg =
-                        std::get<messages::AcknowledgePrevious>(response);
-                    REQUIRE(response_msg.responding_to_id == 123);
-                    REQUIRE(response_msg.with_error ==
-                            errors::ErrorCode::NO_ERROR);
+                    tasks->require_has_ack_for(message);
                 }
             }
         }
@@ -152,20 +109,8 @@ SCENARIO("lid heater task message passing") {
                 AND_THEN("the task should act on the message") {
                     REQUIRE(
                         tasks->get_thermal_plate_queue().backing_deque.empty());
-
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(
-                        std::holds_alternative<messages::AcknowledgePrevious>(
-                            response));
-                    auto response_msg =
-                        std::get<messages::AcknowledgePrevious>(response);
-                    REQUIRE(response_msg.responding_to_id == 555);
-                    REQUIRE(response_msg.with_error ==
-                            errors::ErrorCode::THERMAL_CONSTANT_OUT_OF_RANGE);
+                    tasks->require_has_ack_for(
+                        message, errors::ErrorCode::THERMAL_CONSTANT_OUT_OF_RANGE);
                 }
             }
         }
@@ -178,19 +123,7 @@ SCENARIO("lid heater task message passing") {
             THEN("the task should get the message") {
                 REQUIRE(tasks->get_lid_heater_queue().backing_deque.empty());
                 AND_THEN("the task should respond to the message") {
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(
-                        std::holds_alternative<messages::AcknowledgePrevious>(
-                            response));
-                    auto response_msg =
-                        std::get<messages::AcknowledgePrevious>(response);
-                    REQUIRE(response_msg.responding_to_id == 123);
-                    REQUIRE(response_msg.with_error ==
-                            errors::ErrorCode::NO_ERROR);
+                    tasks->require_has_ack_for(message);
                     AND_WHEN("sending a GetLidTemp query") {
                         auto tempMessage =
                             messages::GetLidTempMessage{.id = 555};
@@ -198,12 +131,9 @@ SCENARIO("lid heater task message passing") {
                             messages::LidHeaterMessage(tempMessage));
                         tasks->run_lid_heater_task();
                         THEN("the response should have the new setpoint") {
-                            REQUIRE(!tasks->get_host_comms_queue()
-                                         .backing_deque.empty());
-                            REQUIRE(std::get<messages::GetLidTempResponse>(
-                                        tasks->get_host_comms_queue()
-                                            .backing_deque.front())
-                                        .set_temp == message.setpoint);
+                            auto response = tasks->get_latest_host_comms_message<
+                                messages::GetLidTempResponse>();
+                            REQUIRE(response.set_temp == message.setpoint);
                         }
                     }
                 }
@@ -228,13 +158,7 @@ SCENARIO("lid heater task message passing") {
                     messages::LidHeaterMessage(tempMessage));
                 tasks->run_lid_heater_task();
                 THEN("the task should respond to the message") {
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    REQUIRE(
-                        std::get<messages::AcknowledgePrevious>(
-                            tasks->get_host_comms_queue().backing_deque.front())
-                            .responding_to_id == 321);
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
+                    tasks->require_has_ack_for(tempMessage);
                     AND_WHEN("sending a GetLidTemp query") {
                         auto tempMessage =
                             messages::GetLidTempMessage{.id = 555};
@@ -242,12 +166,9 @@ SCENARIO("lid heater task message passing") {
                             messages::LidHeaterMessage(tempMessage));
                         tasks->run_lid_heater_task();
                         THEN("the response should have no setpoint") {
-                            REQUIRE(!tasks->get_host_comms_queue()
-                                         .backing_deque.empty());
-                            REQUIRE(std::get<messages::GetLidTempResponse>(
-                                        tasks->get_host_comms_queue()
-                                            .backing_deque.front())
-                                        .set_temp == 0.0F);
+                            auto response = tasks->get_latest_host_comms_message<
+                                messages::GetLidTempResponse>();
+                            REQUIRE(response.set_temp == 0.0F);
                         }
                     }
                 }
@@ -257,14 +178,15 @@ SCENARIO("lid heater task message passing") {
                 tasks->get_host_comms_queue().backing_deque.pop_front();
                 auto tempMessage = messages::DeactivateLidHeatingMessage{
                     .id = 321, .from_system = true};
-                tasks->get_lid_heater_queue().backing_deque.push_back(
-                    messages::LidHeaterMessage(tempMessage));
+                tasks->get_lid_heater_queue().backing_deque.push_back(tempMessage);
                 tasks->run_lid_heater_task();
                 THEN("the task should respond to the message") {
-                    REQUIRE(!tasks->get_system_queue().backing_deque.empty());
-                    REQUIRE(std::get<messages::AcknowledgePrevious>(
-                                tasks->get_system_queue().backing_deque.front())
-                                .responding_to_id == 321);
+                    REQUIRE_FALSE(tasks->get_system_queue().backing_deque.empty());
+                    auto response = tasks->get_system_queue().backing_deque.front();
+                    REQUIRE(std::holds_alternative<messages::AcknowledgePrevious>(response));
+                    auto ack = std::get<messages::AcknowledgePrevious>(response);
+                    REQUIRE(ack.responding_to_id == tempMessage.id);
+                    REQUIRE(ack.with_error == errors::ErrorCode::NO_ERROR);
                 }
             }
             AND_WHEN("sending a DeactivateAll command") {
@@ -274,13 +196,9 @@ SCENARIO("lid heater task message passing") {
                     messages::LidHeaterMessage(tempMessage));
                 tasks->run_lid_heater_task();
                 THEN("the task should respond to the message") {
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    REQUIRE(
-                        std::get<messages::DeactivateAllResponse>(
-                            tasks->get_host_comms_queue().backing_deque.front())
-                            .responding_to_id == 321);
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
+                    auto response = tasks->get_latest_host_comms_message<
+                        messages::DeactivateAllResponse>();
+                    REQUIRE(response.responding_to_id == 321);
                     AND_WHEN("sending a GetLidTemp query") {
                         auto tempMessage =
                             messages::GetLidTempMessage{.id = 555};
@@ -288,12 +206,9 @@ SCENARIO("lid heater task message passing") {
                             messages::LidHeaterMessage(tempMessage));
                         tasks->run_lid_heater_task();
                         THEN("the response should have no setpoint") {
-                            REQUIRE(!tasks->get_host_comms_queue()
-                                         .backing_deque.empty());
-                            REQUIRE(std::get<messages::GetLidTempResponse>(
-                                        tasks->get_host_comms_queue()
-                                            .backing_deque.front())
-                                        .set_temp == 0.0F);
+                            auto gettemp = tasks->get_latest_host_comms_message<
+                                messages::GetLidTempResponse>();
+                            REQUIRE(gettemp.set_temp == 0.0F);
                         }
                     }
                 }
@@ -316,19 +231,7 @@ SCENARIO("lid heater task message passing") {
                     AND_THEN("the task should respond with a busy error") {
                         REQUIRE(tasks->get_thermal_plate_queue()
                                     .backing_deque.empty());
-
-                        REQUIRE(!tasks->get_host_comms_queue()
-                                     .backing_deque.empty());
-                        auto response =
-                            tasks->get_host_comms_queue().backing_deque.front();
-                        tasks->get_host_comms_queue().backing_deque.pop_front();
-                        REQUIRE(std::holds_alternative<
-                                messages::AcknowledgePrevious>(response));
-                        auto response_msg =
-                            std::get<messages::AcknowledgePrevious>(response);
-                        REQUIRE(response_msg.responding_to_id == 808);
-                        REQUIRE(response_msg.with_error ==
-                                errors::ErrorCode::THERMAL_LID_BUSY);
+                        tasks->require_has_ack_for(message, errors::ErrorCode::THERMAL_LID_BUSY);
                     }
                 }
             }
@@ -340,10 +243,8 @@ SCENARIO("lid heater task message passing") {
                 tasks->get_lid_heater_queue().backing_deque.push_back(message);
                 tasks->run_lid_heater_task();
                 THEN("the power is returned correctly") {
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response = std::get<messages::GetLidPowerResponse>(
-                        tasks->get_host_comms_queue().backing_deque.front());
+                    auto response = tasks->get_latest_host_comms_message<
+                        messages::GetLidPowerResponse>();
                     REQUIRE(response.responding_to_id == message.id);
                     REQUIRE_THAT(response.heater,
                                  Catch::Matchers::WithinAbs(0.5, 0.01));
@@ -360,12 +261,9 @@ SCENARIO("lid heater task message passing") {
             messages::LidHeaterMessage(read_message));
         tasks->run_lid_heater_task();
 #if defined(SYSTEM_ALLOW_ASYNC_ERRORS)
-        CHECK(std::holds_alternative<messages::ErrorMessage>(
-            tasks->get_host_comms_queue().backing_deque.front()));
-        auto error_msg = std::get<messages::ErrorMessage>(
-            tasks->get_host_comms_queue().backing_deque.front());
+        auto error_msg = tasks->get_latest_host_comms_message<
+            messages::ErrorMessage>();
         CHECK(error_msg.code == errors::ErrorCode::THERMISTOR_LID_SHORT);
-        tasks->get_host_comms_queue().backing_deque.pop_front();
 #endif
         CHECK(tasks->get_host_comms_queue().backing_deque.empty());
 
@@ -380,22 +278,10 @@ SCENARIO("lid heater task message passing") {
                 AND_THEN("the task should respond with an error") {
                     REQUIRE(
                         tasks->get_lid_heater_queue().backing_deque.empty());
-
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(
-                        std::holds_alternative<messages::AcknowledgePrevious>(
-                            response));
-                    auto response_msg =
-                        std::get<messages::AcknowledgePrevious>(response);
-                    REQUIRE(response_msg.responding_to_id == 124);
-                    REQUIRE(response_msg.with_error !=
-                            errors::ErrorCode::NO_ERROR);
                     REQUIRE(tasks->get_lid_heater_policy().get_heater_power() ==
                             0.0F);
+                    tasks->require_has_ack_for(
+                        message, errors::ErrorCode::THERMISTOR_LID_SHORT);
                 }
             }
         }
@@ -408,19 +294,8 @@ SCENARIO("lid heater task message passing") {
             THEN("the task should get the message") {
                 REQUIRE(tasks->get_lid_heater_queue().backing_deque.empty());
                 AND_THEN("the task should respond with an error") {
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(
-                        std::holds_alternative<messages::AcknowledgePrevious>(
-                            response));
-                    auto response_msg =
-                        std::get<messages::AcknowledgePrevious>(response);
-                    REQUIRE(response_msg.responding_to_id == 123);
-                    REQUIRE(response_msg.with_error !=
-                            errors::ErrorCode::NO_ERROR);
+                    tasks->require_has_ack_for(
+                        message, errors::ErrorCode::THERMISTOR_LID_SHORT);
                     AND_WHEN("sending a GetLidTemp query") {
                         auto tempMessage =
                             messages::GetLidTempMessage{.id = 555};
@@ -428,12 +303,9 @@ SCENARIO("lid heater task message passing") {
                             messages::LidHeaterMessage(tempMessage));
                         tasks->run_lid_heater_task();
                         THEN("the response should have a setpoint of 0") {
-                            REQUIRE(!tasks->get_host_comms_queue()
-                                         .backing_deque.empty());
-                            REQUIRE(std::get<messages::GetLidTempResponse>(
-                                        tasks->get_host_comms_queue()
-                                            .backing_deque.front())
-                                        .set_temp == 0.0F);
+                            auto gettemp = tasks->get_latest_host_comms_message<
+                                messages::GetLidTempResponse>();
+                            REQUIRE(gettemp.set_temp == 0.0F);
                         }
                     }
                 }
@@ -447,16 +319,8 @@ SCENARIO("lid heater task message passing") {
             THEN("the task should get the message") {
                 REQUIRE(tasks->get_lid_heater_queue().backing_deque.empty());
                 AND_THEN("the task should respond with a temp of 0") {
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(
-                        std::holds_alternative<messages::GetLidTempResponse>(
-                            response));
-                    auto gettemp =
-                        std::get<messages::GetLidTempResponse>(response);
+                    auto gettemp = tasks->get_latest_host_comms_message<
+                        messages::GetLidTempResponse>();
                     REQUIRE(gettemp.responding_to_id == message.id);
                     REQUIRE_THAT(gettemp.current_temp,
                                  Catch::Matchers::WithinAbs(0.0, 0.1));
@@ -475,12 +339,9 @@ SCENARIO("lid heater task message passing") {
             messages::LidHeaterMessage(read_message));
         tasks->run_lid_heater_task();
 #if defined(SYSTEM_ALLOW_ASYNC_ERRORS)
-        CHECK(std::holds_alternative<messages::ErrorMessage>(
-            tasks->get_host_comms_queue().backing_deque.front()));
-        auto error_msg = std::get<messages::ErrorMessage>(
-            tasks->get_host_comms_queue().backing_deque.front());
+        auto error_msg = tasks->get_latest_host_comms_message<
+            messages::ErrorMessage>();
         CHECK(error_msg.code == errors::ErrorCode::THERMISTOR_LID_DISCONNECTED);
-        tasks->get_host_comms_queue().backing_deque.pop_front();
 #endif
         CHECK(tasks->get_host_comms_queue().backing_deque.empty());
 
@@ -495,20 +356,8 @@ SCENARIO("lid heater task message passing") {
                 AND_THEN("the task should respond with an error") {
                     REQUIRE(
                         tasks->get_lid_heater_queue().backing_deque.empty());
-
-                    REQUIRE(
-                        !tasks->get_host_comms_queue().backing_deque.empty());
-                    auto response =
-                        tasks->get_host_comms_queue().backing_deque.front();
-                    tasks->get_host_comms_queue().backing_deque.pop_front();
-                    REQUIRE(
-                        std::holds_alternative<messages::AcknowledgePrevious>(
-                            response));
-                    auto response_msg =
-                        std::get<messages::AcknowledgePrevious>(response);
-                    REQUIRE(response_msg.responding_to_id == 124);
-                    REQUIRE(response_msg.with_error !=
-                            errors::ErrorCode::NO_ERROR);
+                    tasks->require_has_ack_for(
+                        message, errors::ErrorCode::THERMISTOR_LID_DISCONNECTED);
                     REQUIRE(tasks->get_lid_heater_policy().get_heater_power() ==
                             0.0F);
                 }
@@ -525,13 +374,7 @@ SCENARIO("lid heater task message passing") {
             REQUIRE(tasks->get_lid_heater_policy().lid_fans_enabled());
         }
         THEN("the message is acked") {
-            REQUIRE(tasks->get_host_comms_queue().has_message());
-            auto reply = tasks->get_host_comms_queue().backing_deque.front();
-            REQUIRE(
-                std::holds_alternative<messages::AcknowledgePrevious>(reply));
-            auto ack = std::get<messages::AcknowledgePrevious>(reply);
-            REQUIRE(ack.responding_to_id == fan_msg.id);
-            REQUIRE(ack.with_error == errors::ErrorCode::NO_ERROR);
+            tasks->require_has_ack_for(fan_msg);
         }
         AND_WHEN("sending another message to disable the fans") {
             fan_msg.id = 456;
@@ -544,19 +387,13 @@ SCENARIO("lid heater task message passing") {
                 REQUIRE(!tasks->get_lid_heater_policy().lid_fans_enabled());
             }
             THEN("the message is acked") {
-                REQUIRE(tasks->get_host_comms_queue().has_message());
-                auto reply =
-                    tasks->get_host_comms_queue().backing_deque.front();
-                REQUIRE(std::holds_alternative<messages::AcknowledgePrevious>(
-                    reply));
-                auto ack = std::get<messages::AcknowledgePrevious>(reply);
-                REQUIRE(ack.responding_to_id == fan_msg.id);
-                REQUIRE(ack.with_error == errors::ErrorCode::NO_ERROR);
+                tasks->require_has_ack_for(fan_msg);
             }
         }
     }
 }
-TEST_CASE("lid heater error flag handling") {
+
+SCENARIO("lid heater error flag handling") {
     uint32_t timestamp = TIME_DELTA;
     GIVEN("a lid heater task with invalid temperatures") {
         auto tasks = TaskBuilder::build();
@@ -573,13 +410,7 @@ TEST_CASE("lid heater error flag handling") {
             REQUIRE(lid_queue.try_send(set_msg));
             tasks->run_lid_heater_task();
             THEN("the response shows an error") {
-                REQUIRE(host_queue.has_message());
-                auto rsp = host_queue.backing_deque.front();
-                REQUIRE(
-                    std::holds_alternative<messages::AcknowledgePrevious>(rsp));
-                auto response = std::get<messages::AcknowledgePrevious>(rsp);
-                REQUIRE(response.responding_to_id == 123);
-                REQUIRE(response.with_error != errors::ErrorCode::NO_ERROR);
+                tasks->require_has_ack_for(set_msg, errors::ErrorCode::THERMISTOR_LID_SHORT);
             }
         }
         WHEN("sending a DeactivateAll message") {
@@ -593,15 +424,7 @@ TEST_CASE("lid heater error flag handling") {
                 REQUIRE(lid_queue.try_send(set_msg));
                 tasks->run_lid_heater_task();
                 THEN("the response shows an error") {
-                    REQUIRE(host_queue.has_message());
-                    auto rsp = host_queue.backing_deque.front();
-                    REQUIRE(
-                        std::holds_alternative<messages::AcknowledgePrevious>(
-                            rsp));
-                    auto response =
-                        std::get<messages::AcknowledgePrevious>(rsp);
-                    REQUIRE(response.responding_to_id == 123);
-                    REQUIRE(response.with_error != errors::ErrorCode::NO_ERROR);
+                    tasks->require_has_ack_for(set_msg, errors::ErrorCode::THERMISTOR_LID_SHORT);
                 }
             }
         }
